+++ app/client_certbot.py
@@ -10,14 +10,17 @@ class CertbotClient():
         self.webroot_path = kwargs.get('webroot_path')
         self.manual_auth_hook = kwargs.get('manual_auth_hook')
         self.manual_cleanup_hook = kwargs.get('manual_cleanup_hook')
+        self.digitalocean_api_key = kwargs.get('digitalocean_api_key')
         self.options = kwargs.get('options', "")
 
-        if self.challenge not in ("http", "dns"):
+        if self.challenge not in ("http", "dns", "dns_digitalocean"):
             raise Exception('required argument "challenge" not set.')
         if self.challenge == "http" and self.webroot_path is None:
             raise Exception('required argument "webroot_path" not set. Required when using challenge "http"')
         if self.challenge == "dns" and (self.manual_auth_hook is None or self.manual_cleanup_hook is None):
             raise Exception('required argument "manual_auth_hook" or "manual_manual_hook" not set. Required when using challenge "dns"')
+        if self.challenge == "dns_digitalocean" and self.digitalocean_api_key is None:
+            raise Exception('required argument "digitalocean_api_key" not set. Required when using challenge "dns_digitalocean"')
 
 
     def run(self, cmd):
@@ -58,6 +61,10 @@ class CertbotClient():
         if self.challenge == 'dns':
             c = "--manual --manual-public-ip-logging-ok --preferred-challenges dns --manual-auth-hook {} --manual-cleanup-hook {}".format(self.manual_auth_hook, self.manual_cleanup_hook)
 
+        if self.challenge == 'dns_digitalocean':
+            c = "--dns-digitalocean --dns-digitalocean-credentials ~/secrets/digitalocean.ini --dns-digitalocean-propagation-seconds 60"
+            self.run( """echo dns_digitalocean_token ={}>~/secrets/digitalocean.ini""".format(self.digitalocean_api_key))
+
         output, error, code = self.run("""certbot certonly \
                     --agree-tos \
                     --domains {domains} \
@@ -75,6 +82,9 @@ class CertbotClient():
 
         ret_error = False
         ret_created = True
+
+        if self.challenge == 'dns_digitalocean':
+            self.run("""rm ~/secrets/digitalocean.ini""")
 
         if b'urn:acme:error:unauthorized' in error:
             logger.error('Error during ACME challenge, is the domain name associated with the right IP ?')