+++ app/hooks/do/manual-auth-hook.sh
@@ -1,2 +1,0 @@
-DIR="$(dirname $0)"
-python $DIR/manual-auth-hook.py