+++ app/hooks/do/manual-cleanup-hook.sh
@@ -1,2 +1,0 @@
-DIR="$(dirname $0)"
-python $DIR/manual-cleanup-hook.py